backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-docx>=1.0.1
langdetect>=1.0.9
python-dotenv>=1.0.0
pydantic>=2.4.2
langchain_groq>=0.2.5